Upgrade to {fmt} 8.1.1 and fix compilation errors
The newest version of fmt doesn't allow the printing of enum class
variables. This is how it's meant to be so we were wrong :( Use
static_cast for the moment just to get things working again.

+++ src/nuclide.cpp
@@ -557,7 +557,7 @@ std::string Nuclide::writeAsEPS(const Options& draw) const
                      "{} {} {} {} {} {} curve Nucleus\n",
                      A,
                      symbol,
-                     getDisplayMode(draw.chart_colour),
+                     static_cast<int>(getDisplayMode(draw.chart_colour)),
                      getIsotopeTextColour(draw.chart_colour, draw.write_isotope),
                      draw.write_isotope ? fmt::format("({}) ({})", symbol, A) : "",
                      colour,
